phases/00-setup-and-tooling/04-apis-and-keys/pyproject.toml
[project]
name = "04-apis-and-keys"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "anthropic>=0.101.0",
    "openai>=2.36.0",
    "python-dotenv>=1.2.2",
]
